requirements.txt
# ScriptGen requirements

# Core dependencies
sqlalchemy>=2.0.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
json5>=0.9.14
tenacity>=8.2.3

# Frontend dependencies
streamlit>=1.24.0
streamlit-extras>=0.3.0
watchdog>=3.0.0

# Database
sqlalchemy>=1.4.0

# LLM APIs
mistralai>=0.0.6

# Utilities
absl-py==2.1.0
accelerate==0.33.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
altair==5.4.0
annotated-types==0.7.0
anthropic==0.34.1
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asteroid-filterbanks==0.4.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
audioread==3.0.1
av==12.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
braceexpand==0.1.7
cachetools==5.5.0
cairocffi==1.7.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
contourpy==1.2.1
ctranslate2==4.3.1
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==4.4.2
defusedxml==0.7.1
diffusers==0.30.2
distro==1.9.0
docopt==0.6.2
editdistance==0.8.1
einops==0.8.0
executing==2.1.0
faster-whisper==1.0.3
fastjsonschema==2.20.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
gizeh==0.1.11
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
google-pasta==0.2.0
googleapis-common-protos==1.64.0
greenlet==3.0.3
groq==0.10.0
grpcio==1.65.5
grpcio-status==1.62.3
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
inflect==7.4.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
jax==0.4.31
jaxlib==0.4.31
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jiwer==3.0.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
julius==0.2.7
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.5.0
kiwisolver==1.4.5
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.35
langchain-openai==0.1.23
langchain-text-splitters==0.2.2
langsmith==0.1.104
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.7
llvmlite==0.43.0
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.0.2
ml-dtypes==0.3.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.4.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nemo-toolkit==1.21.0
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.23.5
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.2
opcai==1.43.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
optuna==4.0.0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
primePy==1.3
proglog==0.1.10
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.audio==3.3.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytorch-metric-learning==2.6.0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
rapidfuzz==3.9.7
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
retina-face==0.0.17
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
semver==3.0.2
Send2Trash==1.8.3
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.4.0
speechbrain==1.0.1
streamlit==1.37.1
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
tf_keras==2.16.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
torch==2.2.2
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.2
torchmetrics==1.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.0
typeguard==4.3.0
typer==0.12.5
types-python-dateutil==2.9.0.20240821
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.8.0
webdataset==0.2.100
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wget==3.2
widgetsnbextension==4.0.13
wrapt==1.16.0
yarl==1.9.4
youtokentome==1.0.6
zipp==3.20.1
kokoro>=0.3.4
openai-whisper